app/services/ma_data_collectors/bam_collector.py
import logging
from datetime import datetime
from typing import Dict, Optional
import requests
from bs4 import BeautifulSoup
from .base_collector import BaseCollector

logger = logging.getLogger(__name__)

class BAMCollector(BaseCollector):
    """Collecteur de données de Bank Al-Maghrib."""
    
    def __init__(self, use_cache: bool = True):
        """Initialise le collecteur BAM.
        
        Args:
            use_cache: Si True, utilise le cache pour les requêtes
        """
        super().__init__(use_cache)
        self.base_url = "https://www.bam.ma"
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
            'Accept': 'application/json, text/html, application/xhtml+xml',
            'Accept-Language': 'fr,fr-FR;q=0.9,en;q=0.8,ar;q=0.7',
            'Referer': self.base_url,
        })

    def get_indicators(self) -> Dict:
        """Récupère les principaux indicateurs de la BAM."""
        if self.use_cache and 'bam_indicators' in self.cache:
            return self.cache['bam_indicators']
            
        try:
            # En production, vous pourriez implémenter le vrai scraping ici
            # Pour l'instant, nous utilisons des données simulées
            data = {
                "date": datetime.now().strftime("%Y-%m-%d"),
                "source": "BAM",
                "indicators": {
                    "tmm": {
                        "value": 1.5, 
                        "unit": "%", 
                        "description": "Taux moyen des marchés monétaires",
                        "category": "monetary"
                    },
                    "taux_change": {
                        "value": 10.5, 
                        "unit": "MAD/EUR", 
                        "description": "Taux de change EUR/MAD",
                        "category": "forex"
                    },
                    "reserves": {
                        "value": 350, 
                        "unit": "Milliards MAD", 
                        "description": "Réserves de change",
                        "category": "forex"
                    },
                    "taux_directeur": {
                        "value": 3.0, 
                        "unit": "%", 
                        "description": "Taux directeur BAM",
                        "category": "monetary"
                    }
                }
            }
            
            if self.use_cache:
                self.cache['bam_indicators'] = data
            return data
            
        except Exception as e:
            logger.error(f"Erreur lors de la récupération des données BAM: {e}")
            return {}
            
    def collect(self) -> Dict:
        """Méthode principale de collecte des données.
        
        Returns:
            Dict: Données formatées pour le stockage en base
        """
        indicators = self.get_indicators()
        
        # Formatage des données pour correspondre au modèle de données
        formatted_data = {
            "date": indicators.get("date"),
            "source": indicators.get("source"),
            "indicators": {}
        }
        
        # Mapping des indicateurs au format attendu
        for key, indicator in indicators.get("indicators", {}).items():
            formatted_data["indicators"][f"bam_{key}"] = {
                "value": indicator["value"],
                "unit": indicator["unit"],
                "description": indicator["description"],
                "category": indicator["category"]
            }
            
        return formatted_data