requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Caching==2.1.0
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Assets==2.1.0

# Database
SQLAlchemy==2.0.20
alembic==1.12.1
psycopg2-binary==2.9.9  # Pour PostgreSQL

# Authentication & Security
PyJWT==2.8.0
email-validator==2.0.0
passlib==1.7.4

# Data Processing
pandas==2.1.0
numpy==1.26.0
python-dateutil==2.8.2
pytz==2023.3

# Web & API
requests==2.31.0
requests-html==0.10.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Dashboard
plotly==5.18.0
dash==2.14.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0

# Background Tasks
APScheduler==3.10.4
tzlocal==5.0.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
webassets==2.1.0
cssmin==0.2.0
jsmin==3.0.1

# Development
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
